Move database tests (#7318)
